server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
RPi.GPIO==0.7.1